animator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
boto3==1.34.150
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.10.0.84
librosa==0.10.1
scipy==1.11.4
# GPU build (CUDA 11.8). For CPU, comment these 3 and install CPU wheels instead.
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
